backend/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.0
psycopg2-binary==2.9.6
firebase-admin==6.1.0
twilio==7.18.0
